+++ CHANGELOG.md
@@ -64,3 +64,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Added cursor position tracking for the ORCA output file. The TOC view will automatically highlight and expand the corresponding TOC entry when the user is navigating through the ORCA output file based on the current line number. The highlight and expand status of the TOC view will be updated when the user is moving the cursor to a different line in the ORCA output file.
 - Added support for changing the highlight symbols for the TOC entries.
 - Added support for saving the collapsed status of the TOC view in the settings of the extension. When navigating to different ORCA output files, the TOC view status for each file will be restored.
+
+## [0.4.1] - 2023-12-30
+
+- Added command `orca_toc: Toggle Collapse/Expand All TOC Entries` to toggle collapse/expand all TOC entries at once. The collapsed status of the TOC view will be preserved when navigating between different ORCA output files. The default collapsed status of the TOC view can be changed in the settings of the extension.
+- Update of the `patterns.json` file for better parsing of the ORCA output file.